ENTREGABLES/00_POYECTO_ML/requirements.txt
absl-py==0.13.0
altair==4.1.0
argon2-cffi==20.1.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
base58==2.1.0
beautifulsoup4==4.9.3
bleach==3.3.0
blinker==1.4
branca==0.4.2
Brotli==1.0.9
cached-property==1.5.2
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cycler==0.10.0
Cython==0.29.21
dash==1.20.0
dash-bootstrap-components==0.12.2
dash-colorscales==0.0.4
dash-core-components==1.16.0
dash-daq==0.5.0
dash-html-components==1.1.3
dash-renderer==1.9.1
dash-table==4.11.3
debugpy==1.3.0
decorator==4.4.2
defusedxml==0.7.1
dnspython==1.16.0
dtale==1.47.0
entrypoints==0.3
et-xmlfile==1.1.0
Flask==2.0.0
Flask-Compress==1.9.0
flask-ngrok==0.0.25
flatbuffers==1.12
folium==0.12.1
future==0.18.2
gast==0.4.0
gensim==4.0.1
geocoder==1.38.1
geographiclib==1.50
geopy==2.1.0
gitdb==4.0.7
GitPython==3.1.17
google-auth==1.32.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
greenlet==1.1.0
grpcio==1.34.1
h5py==3.1.0
idna==2.10
imageio==2.9.0
importlib-metadata==3.10.1
ipykernel==6.0.1
ipython==7.25.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.0
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.2.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kaggle==1.5.12
kaleido==0.2.1
Keras==2.4.3
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lxml==4.6.3
lz4==3.1.3
Markdown==3.3.4
MarkupSafe==2.0.0
matplotlib==3.4.1
matplotlib-inline==0.1.2
matplotlip==0.2
missingno==0.4.2
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
nltk==3.6.2
notebook==6.4.0
numpy==1.19.5
oauthlib==3.1.0
opencv-contrib-python==4.5.2.54
openpyxl==3.0.7
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
patsy==0.5.1
pickleshare==0.7.5
Pillow==8.2.0
plotly==4.14.3
ppscore==1.2.0
prometheus-client==0.10.1
prompt-toolkit==3.0.18
protobuf==3.17.0
pyarrow==4.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydeck==0.6.2
Pygments==2.8.1
pymongo==3.11.4
PyMySQL==1.0.2
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.1
python-slugify==5.0.2
pytz==2021.1
PyWavelets==1.1.1
pywin32==300
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==22.0.3
ratelim==0.1.6
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.7.2
scikit-image==0.18.1
scikit-learn==0.24.2
scipy==1.6.3
seaborn==0.11.1
selenium==3.141.0
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
smart-open==5.1.0
smmap==4.0.0
soupsieve==2.2.1
SQLAlchemy==1.4.15
squarify==0.4.3
statsmodels==0.12.2
streamlit==0.82.0
strsimpy==0.2.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
terminado==0.9.5
testpath==0.5.0
text-unidecode==1.3
threadpoolctl==2.1.0
tifffile==2021.4.8
toml==0.10.2
toolz==0.11.1
tornado==6.1
tqdm==4.60.0
traitlets==5.0.5
tweepy==3.10.0
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.4
utils==1.0.1
validators==0.18.2
vega-datasets==0.9.0
watchdog==2.1.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xarray==0.18.2
xgboost==1.4.2
xlrd==2.0.1
zipp==3.4.1
